Add index files for dataset folders
Basically just rename old index -> table-data
and add exports for all files to index.
